WIP Enable Chrome Runtime
The cookie directories for each service integration must be moved into
the root config directory, as subdirectories in their current form
(obs_profile_cookies/<cookie_id>) are not supported. This is
intentional, and according to Marshall this was never a supported setup.
cache_path *must* be a direct child of the root_cache_path. Invalid
cache_path_ is silently treated as Incognito Mode and cookies are not
stored.

TODO: Disable ability to set browser as default
TODO: Disable ability to switch profiles (as it currently crashes)